requirements.txt
langchain_classic==1.0.4
langchain_community==0.4.1
langchain_core==1.3.0
langchain_google_genai==4.2.2
langchain_ollama==1.1.0
langchain_text_splitters==1.1.2
protobuf==7.34.1
python-dotenv==1.2.2
